componentized remaining pages
