+++ pages/tabroom.tsx
@@ -1,43 +1,11 @@
 /* eslint-disable @next/next/no-img-element */
 import type { NextPage } from 'next';
-import Link from 'next/link';
 import React, { FunctionComponent,  useState } from 'react';
 import Head from 'next/head';
 import styles from '../styles/Q.module.css';
-import type { Tool, Team, Speaker } from './typedeclarations';
 import NavigationBar from '../components/nav/NavigationMenu';
-
-const BreakTable: FunctionComponent<{teams: Array<Team>, division: string, topx: number}> = ({teams, division, topx}) => {
-  return (
-    <>
-      {(teams.filter(e => e.division==division)).slice(0,topx).map((item, index) => (
-        <tr key={index} style={{height: "1rem"}}>
-          <td style={{width:"10%"}}>{index+1}</td><td style={{width:"10%"}}>{item.id}</td><td style={{width:"40%"}}>{item.speaker1.name_cn} - {item.speaker1.school}</td><td style={{width:"40%"}}>{item.speaker2.name_cn} - {item.speaker2.school}</td>
-        </tr>
-      ))}
-    </>
-  )
-}
-
-const SpeakerTable: FunctionComponent<{speakers: Array<Speaker>, division: string, topx: number}> = ({speakers, division, topx}) => {
-  return (
-    <>
-      {(speakers.filter(e => e.division==division)).slice(0,topx).map((item, index) => (
-        <tr key={index} style={{height: "1rem"}}>
-          <td style={{width:"10%"}}>{index+1}</td><td style={{width:"10%"}}>{item.id}</td><td style={{width:"40%"}}>{item.name_cn}</td><td style={{width:"40%"}}>{item.school}</td>
-        </tr>
-      ))}
-    </>
-  )
-}
-
-const toolbox: Tool[] = [
-  {id: 1, name: "Pairings Generator", description: "Generate pairings image from horizontal schematic", link: "/pair", active: false},
-  {id: 2, name: "Results Image Generator", description: "Generator results as an image", link: "/results", active: true},
-  {id: 3, name: "Results Spreadsheet Generator", description: "Generate results for a given division as a csv file", link: "/resultscsv", active: false},
-  {id: 10, name: "Tabroom Import Spreadsheet Convertor", description: "Convert DLC namelist to Tabroom format spreadsheet", link: "/tabroom", active: false},
-  {id: 99, name: "Evaluate Judges", description: "Judge evaluation system", link: "/evaluate", active: false}
-];
+import Speaker from '../types/Speaker';
+import Team from '../types/Team';
 
 const Home: NextPage = () => {
   const readFile = async (file: any) => {
