+++ types/Debate.ts
@@ -1,0 +1,9 @@
+export type Debate = {
+    flight: string;
+    teamA: string;
+    teamB: string;
+    roomCode: string;
+    judgeName: any;
+};
+
+export default Debate;
