+++ types/Speaker.ts
@@ -1,0 +1,10 @@
+export type Speaker = {
+    division: string,
+    id: number,
+    teamid: number
+    name_cn: string,
+    name_en: string,
+    school: string
+};
+
+export default Speaker;
