+++ types/Team.ts
@@ -1,0 +1,10 @@
+import Speaker from "./Speaker";
+
+export type Team = {
+    division: string,
+    id: number,
+    speaker1: Speaker,
+    speaker2: Speaker
+};
+
+export default Team;
